paddlemix/examples/jina_clip_v2/requirement.txt
pillow
tqdm
paddlenlp==3.0.0b2
einops 
timm